+++ bundles/framework/search/view/DefaultSearchView.js
@@ -383,7 +383,7 @@ Oskari.clazz.define(
             }
             sandbox.request(
                 me.instance.getName(),
-                moveReqBuilder(result.lon, result.lat, zoom, false)
+                moveReqBuilder(result.lon, result.lat, zoom)
             );
 
             var loc = this.instance.getLocalization('resultBox'),