+++ bundles/mapping/mapmodule/AbstractMapModule.js
@@ -887,7 +887,7 @@ Oskari.clazz.define(
 
             var lonlat = this.getMapCenter();
             this.updateDomain();
-            var evt = sandbox.getEventBuilder('AfterMapMoveEvent')(lonlat.lon, lonlat.lat, this.getMapZoom(), false, this.getMapScale(), creator);
+            var evt = sandbox.getEventBuilder('AfterMapMoveEvent')(lonlat.lon, lonlat.lat, this.getMapZoom(), this.getMapScale(), creator);
             sandbox.notifyAll(evt);
         },
 /* --------------- /MAP STATE ------------------------ */