+++ packages/framework/bundle/event-map/bundle.js
@@ -63,25 +63,10 @@ function() {
 
 		{
 			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/event/common/features-available-event.js"
-		}, {
-			"type" : "text/javascript",
 			"src" : "../../../../sources/framework/event/common/after-map-layer-add-event.js"
 		}, {
 			"type" : "text/javascript",
 			"src" : "../../../../sources/framework/event/common/after-map-layer-remove-event.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/event/common/after-map-move-event.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/event/common/after-map-move-start-event.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/event/common/after-show-map-layer-info-event.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/event/common/mouse-hover-event.js"
 		}, {
 			"type" : "text/javascript",
 			"src" : "../../../../sources/framework/event/common/MapLayerEvent.js"